store: Append to existing writes more aggressively
So far, new write requests were appended to existing ones only
opportunistically; this commit makes it so that we actually hold back
writing to allow processing to append to an existing write request.

Only write a batch if its size is beyond WRITE_BATCH_SIZE or if it is older
than WRITE_BATCH_DURATION. Since we don't have a way to wait on the batch
growing big enough, poll for that every 2s

This also requires that stopping or flushing the WritableStore sets
`batch_writes` to `false`, mostly for tests, to ensure that all pending
writes get written out before the background writer shuts down.