graph, store: Calculate entity count change from in-memory data
When changes to entities get combined, inserting a row can either create a
new entity, or update an existing version, and the number of rows inserted
is no longer an accurate count of new entities.

Instead of relying on data coming from the database, we now use the
in-memory representation of changes to determine how applying a batch
changes the number of entities in a deployment.